+++ AddOns/TidyPlates_ThreatPlates/Commands.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 local L = Addon.ThreatPlates.L
 
-Addon.DEBUG = Addon.ThreatPlates.Meta("version") == "12.2.1"
+Addon.DEBUG = Addon.ThreatPlates.Meta("version") == "12.2.3"
 
 local function toggleDPS()
 	if Addon.db.profile.optionRoleDetectionAutomatic then