+++ AddOns/TidyPlates_ThreatPlates/CHANGELOG.md
@@ -1,3 +1,5 @@
-# 12.2.1 (2025-06-01)
+# 12.2.3 (2025-06-18)
 
-* Fixed a bug that prevented nameplates from reappearing after a Hunter's Feign Death had ended [GH-583].
+* Fixed a bug in the script editor that caused updates to script code to be applied to the wrong custom style.
+* Fixed a bug that prevented some nameplate UI elements from being updated when script code was updated.
+* Updated TOC version for Patch 11.1.7 for WoW Mainline.